% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_composition.R
\name{Community_Composition}
\alias{Community_Composition}
\title{Summarize number or percent of species by PlotKey and provided grouping variables.}
\usage{
Community_Composition(SpeciesList, method = "percent", tall = F, ...)
}
\arguments{
\item{SpeciesList}{List of species by plot that will be summarized, including the grouping variables
for which data is being summarized by.}

\item{method}{character string. The method used for the produced summary table. Can
be \code{"percent"}, \code{"mean"}, or {"count"}. Defaults to \code{"percent"}.}

\item{tall}{indicates whether output will be tall table or wide. Defaults to FALSE.}

\item{...}{grouping variables from the master species list to be used in calculating list.}
}
\description{
Basic function that takes species list and summarizes either the percent or count
of species in that group. List provided to function should already be cleaned to remove
duplicate species or unwanted grouping variable categories.
}
