% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_metrics.R
\name{Community_Richness}
\alias{Community_Richness}
\title{Functions for summarizing community makeup for specific grouping variables.}
\usage{
Community_Richness(
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory"
)
}
\arguments{
\item{SpeciesList}{List of species by PlotKey that will be summarized. Can be either spp_inventory
produced by \code{gather_spp_inventory_lentic} or \code{gather_lpi_lentic}.}

\item{masterspecieslist}{Dataframe containing all possible species codes with
necessary grouping variable information.}

\item{listtype}{character string. Indicates the source of the SpeciesList provided. Can either be
\code{"speciesinventory"} or \code{"lpi"}. Defaults to speciesinventory.}

\item{method}{character string. The method used for the produced summary table. Can
be \code{"percent"}, \code{"mean"}, or {"count"}.}
}
\description{
Functions for summarizing community makeup for specific grouping variables.
}
