% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{pct_FoliarCover}
\alias{pct_FoliarCover}
\alias{pct_BasalCover}
\alias{pct_TotalAbsoluteCover}
\alias{pct_NativeCover}
\alias{pct_NoxiousCover}
\alias{pct_HydroNoFACCover}
\alias{pct_HydroWithFACCover}
\alias{pct_GrowthHabitSubCover}
\alias{pct_GrowthHabitCover}
\alias{pct_DurationCover}
\alias{pct_DurationGrowthHabitCover}
\alias{pct_DurationGrowthHabitSubCover}
\alias{pct_NativeGrowthHabitSubCover}
\alias{pct_NativeGrowthHabitCover}
\alias{pct_NativeDurationGrowthHabitSubCover}
\alias{pct_StabilityClassCover}
\alias{pct_SGGroupCover}
\alias{pct_NonPlantGroundCover}
\alias{pct_UnknownCover}
\alias{pct_FunctionalGroupCover}
\alias{pct_AbsoluteSpeciesCover}
\title{Calculate percent cover metrics from LPI}
\usage{
pct_FoliarCover(lpi_tall, nationalspecieslist = NULL, unit = "by_plot")

pct_BasalCover(lpi_tall, unit = "by_plot")

pct_TotalAbsoluteCover(lpi_tall, unit = "by_plot")

pct_NativeCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_NoxiousCover(
  header,
  lpi_tall,
  statespecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_HydroNoFACCover(
  header,
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_HydroWithFACCover(
  header,
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_GrowthHabitSubCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_GrowthHabitCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_DurationCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_DurationGrowthHabitCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_DurationGrowthHabitSubCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_NativeGrowthHabitSubCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_NativeGrowthHabitCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_NativeDurationGrowthHabitSubCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unknowncodes,
  unit = "by_plot"
)

pct_StabilityClassCover(
  header,
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_SGGroupCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "absolute",
  unit = "by_plot"
)

pct_NonPlantGroundCover(
  lpi_tall,
  hit = "any",
  nationalspecieslist,
  unit = "by_plot"
)

pct_UnknownCover(
  lpi_tall,
  nationalspecieslist,
  covertype = "relative",
  unit = "by_plot"
)

pct_FunctionalGroupCover(lpi_tall, nationalspecieslist, covertype = "absolute")

pct_AbsoluteSpeciesCover(lpi_tall, nationalspecieslist, unit = "by_plot")
}
\arguments{
\item{lpi_tall}{A tall/long-format data frame. Use the data frame from the \code{gather_lpi_lentic()} output.}

\item{nationalspecieslist}{Data frame. The centrally managed master species list should be used. The assumed structure is that each row is unique on its Symbol.}

\item{unit}{String. The sampling unit by which data should be summarized. Should be \code{by_plot}, \code{by_line} or \code{by_geosurface} (for data from Lotic-Integration Plots). Defaults to \code{by_plot}.}

\item{covertype}{Character string. "relative" or "absolute". Specifies the kind of cover calculation. Relative cover is only used for calculations on vascular plant species and specifies the percent of overall hits made up of a particular species or group. Absolute cover is the percent of the pin drops made up by a particular species or group.}

\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output. Used in Noxious and Wetland Indicator calculations to specify the plot region or state.}

\item{statespecieslist}{Data frame. The centrally managed master species list should be used. This dataframe should contain a unique record for each Symbol-SpeciesState combination.}

\item{unknowncodes}{Optional data frame. Use the data frame from the \code{gather_unknowns_lentic()} output. Unknown species list matching unknown codes to their duration and Growth habit. This is used to fill in duration and growth habit for plants in LPI never identified to a species or genus with those fields specified. If argument is unused, all unknown species without Duration or Growth Habit specified will be filtered out before being passed on to \code{pct_cover_lentic()}.}

\item{hit}{Character string. "any", "first" or "basal". Only used in \code{pct_NonPlantGroundCover()}, where relative cover is not calculated. If "any" is used, any layer will be used to calculate non-plant cover. If "first" is used, only \code{"TopCanopy"} hits will be counted. If "basal" is used, only \code{"SoilSurface"} hits will be counted. Defaults to "any".}
}
\value{
Wide data frame of percent cover by plot of different categories.
}
\description{
Metric calculation functions for specific grouping variables. These functions perform various modifications to the \code{lpi_tall} data frame in preparation for the \code{pct_cover_lentic()} function, then filters the resulting data frame to the categories of interest.

Some functions pull information from a variety of tables from the field season data used to categorize variables. Region- or state-specific categorization requires joining the header data frame to the lpi_tall data frame. Growth Form and Duration calculations can optionally pull in data for unknowns entered by crews in the Unknown Plant form. Wetland Indicator Status and Noxious species group categories together before pushing to the \code{pct_cover_lentic()} function.

Relative and absolute cover calculations also require different handling of plants identified to genus or higher taxonomic or plant groupings due to the way the percent denominators are calculated. Relative cover calculations remove all species without relevant categorization to ensure only plant hits with an assigned trait are included in calculations. (Example: In calculating relative native cover, plant hits of Elymus species are filtered out of the dataset before using the \code{pct_cover_lentic()} function.)

In contrast, absolute cover calculations submit all hits to the \code{pct_cover_lentic()} function, as these will be used to calculate the number of pin drops (i.e. the denominator). Absolute cover calculations filter out uncategorized hits after cover has been calculated.

Absolute and Relative cover metrics can be calculated all at once using \code{CombineRelativeCoverMetrics} and \code{CombineAbsoluteCoverMetrics}.
}
