% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance_index.R
\name{disturbance_index}
\alias{disturbance_index}
\title{Calculate disturbance index from Human and Natural Disturbances form}
\usage{
disturbance_index(disturbances)
}
\arguments{
\item{disturbances}{A tall/long-format data frame. Use the data frame from the \code{pct_AbsoluteSpeciesCover} output, then join desired species traits.}
}
\description{
Function used to calculate plot-level and landscape-level disturbance index. Function uses the scope and impact scores to calculate a weighted score for each observed disturbance, then sums those scores together for an overall plot score.
}
