% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominancetest.R
\name{dominance_test}
\alias{dominance_test}
\title{Calculate whether each site passes the Dominance Test and Prevalence Test}
\usage{
dominance_test(header, lpi_tall, masterspecieslist, test)
}
\arguments{
\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output. Used in Noxious and
Wetland Indicator calculations to specify the plot region or state.}

\item{lpi_tall}{A tall/long-format data frame. Use the data frame from the \code{gather_lpi_lentic()} output.}

\item{masterspecieslist}{Data frame. The centrally managed master species list should be used.}

\item{bystrata}{logical. Indicates whether Dominance Test should be performed by strata, as is recommended by the USACE.}
}
\value{
Dataframe showing site scores on several ways to classify dominance of hydrophytic species.
}
\description{

}
