% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_lentic.R
\name{gather_lentic}
\alias{gather_lentic}
\alias{gather_lpi_lentic}
\alias{gather_species_inventory_lentic}
\alias{gather_unknowns_lentic}
\alias{gather_height_lentic}
\alias{gather_annualuse}
\alias{gather_woodyspecies}
\alias{gather_hummocks}
\title{Species gathering functions}
\usage{
gather_lpi_lentic(dsn)

gather_species_inventory_lentic(dsn)

gather_unknowns_lentic(dsn)

gather_height_lentic(dsn)

gather_annualuse(dsn)

gather_woodyspecies(dsn)

gather_hummocks(dsn)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extensions) of the geodatabase containing the table of interest.}
}
\value{
Data frame containing the data from a lentic AIM form
}
\description{
This group of functions allow you to transform Survey123 data uploaded to
AGOL from wide format to long format. It contains six functions for transforming Species Richness,
Unknown Plants, LPI, Heights from LPI, and Woody Species detail tables.
}
