% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_lpi_lentic.R
\name{gather_lpi}
\alias{gather_lpi}
\alias{gather_lpi_lentic}
\title{Gather LPI data from Lentic FGDB.}
\usage{
gather_lpi_lentic(dsn)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}
}
\value{
A data frames containing the data from the LPI pin intercepts.
}
\description{
Given a list of data frames containing tblLPIHeader and tblLPIDetail, create a tall format data frame for canopy data from LPI and one for heights from the specialized height fields.
}
\concept{<gather>}
