% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_riparianwetland.R
\name{gather_RiparianWetland}
\alias{gather_RiparianWetland}
\alias{gather_lpi_lentic}
\alias{gather_species_inventory_lentic}
\alias{gather_unknowns_lentic}
\alias{gather_height_lentic}
\alias{gather_annualuse}
\alias{gather_woodyspecies}
\alias{gather_hummocks}
\title{Pivoting functions to transform raw data tables from wide to tall format for use in indicator calculations.}
\usage{
gather_lpi_lentic(dsn)

gather_species_inventory_lentic(dsn)

gather_unknowns_lentic(dsn)

gather_height_lentic(dsn)

gather_annualuse(dsn)

gather_woodyspecies(dsn)

gather_hummocks(dsn)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extensions) of the geodatabase containing the table of interest.}
}
\value{
tall Data frame containing the data from a detail table from a Riparian and Wetland AIM file geodatabase.
}
\description{
This group of functions allow you to transform Survey123 data uploaded to
AGOL from wide format to tall format. It contains seven functions for transforming species inventory,
unknown plants, LPI, heights from LPI and woody species, woody species, annual use, and hummocks detail tables.
}
