% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdbconversion.R
\name{gdbconversion}
\alias{gdbconversion}
\title{Convert FGDB into separate .csv files}
\usage{
gdbconversion(dsn, RawDataFolder, filetype = "csv")
}
\arguments{
\item{dsn}{location of the File Geodatabase}

\item{RawDataFolder}{location where all files will be stored. If exporting to csv, this should be a preexisting folder where all csvs will be placed. If exporting to geopackage, the location must end in '.gpkg' and the file will be created by the script.}

\item{filetype}{String. Export format. Can be \code{csv} or \code{gpkg}. Defaults to \code{csv}.}
}
\description{
Function converts all separate tables found in a FGDB into separate csv files to be more accessible to others. Some schema changes are accomplished to ensure output is interpretable outside of a relational geodatabase format.
}
