% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_metrics.R
\name{pct_FoliarCover}
\alias{pct_FoliarCover}
\alias{pct_BasalCover}
\alias{pct_TotalAbsoluteCover}
\alias{pct_NativeCover}
\alias{pct_NoxiousCover}
\alias{pct_HydrophyteCover}
\alias{pct_HydroFACCover}
\alias{pct_GrowthHabitCover}
\alias{pct_DurationCover}
\alias{pct_NonPlantGroundCover}
\alias{CombineRelativeCoverMetrics}
\alias{CombineAbsoluteCoverMetrics}
\title{Calculate percent cover from LPI}
\usage{
pct_FoliarCover(lpi_tall)

pct_BasalCover(lpi_tall)

pct_TotalAbsoluteCover(lpi_tall)

pct_NativeCover(lpi_tall, masterspecieslist, covertype = "relative")

pct_NoxiousCover(header, lpi_tall, masterspecieslist, covertype = "absolute")

pct_HydrophyteCover(
  header,
  lpi_tall,
  masterspecieslist,
  covertype = "relative"
)

pct_HydroFACCover(header, lpi_tall, masterspecieslist, covertype = "relative")

pct_GrowthHabitCover(
  lpi_tall,
  masterspecieslist,
  unknowncodelist,
  covertype = "relative"
)

pct_DurationCover(
  lpi_tall,
  masterspecieslist,
  unknowncodelist,
  covertype = "relative"
)

pct_NonPlantGroundCover(lpi_tall, hit = "any")

CombineRelativeCoverMetrics(
  header,
  lpi_tall,
  masterspecieslist,
  unknowncodelist
)

CombineAbsoluteCoverMetrics(
  header,
  lpi_tall,
  masterspecieslist,
  unknowncodelist
)
}
\arguments{
\item{lpi_tall}{source of lpi_tall data frame.}

\item{masterspecieslist}{Character string. Full file path (including extension) to the file
containing the species list.}

\item{covertype}{Character string. "relative" or "absolute". Specifies the kind of cover calculation.
Relative cover is only used for calculations on vascular plant species and specifies the percent of
overall hits made up of a particular species or group. Absolute cover is the percent of the pin
drops made up by a particular species or group.}

\item{header}{source of header data frame.}

\item{unknowncodelist}{optional dataframe. Unknown species list matching unknown codes to their duration and
Growth habit. This is used to fill in duration and growth habit for plants in LPI never identified to a
species or genus with those fields specified. If argument is unused, all unknown species without duration or
Growth Habit specified will be filtered out before being passed on to pct_cover_lentic.}
}
\value{
Dataframe of specified percent cover of different categories by plot.
}
\description{
Cover metrics return a \code{data.frame} of percent cover of specified categories calculated from
gathered LPI data across plots. These defined functions execute the most common cover categories calculated
from LPI.
}
