% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_byspecies.R
\name{metrics_byspecies}
\alias{metrics_byspecies}
\title{Combine species-level indicator information into one data.frame.}
\usage{
metrics_byspecies(
  header,
  spp_inventory,
  lpi_tall,
  height_tall,
  woody_tall,
  annualuse_tall,
  masterspecieslist
)
}
\arguments{
\item{header}{data.frame produced by \code{gather_header_lentic}.}

\item{spp_inventory}{data.frame produced by \code{gather_spp_inventory_lentic}.}

\item{lpi_tall}{tall data.frame of LPI detail data produced by \code{gather_lpi_lentic}}

\item{height_tall}{tall data.frame of height data.}

\item{woody_tall}{tall data.frame of woody use data.}

\item{annualuse_tall}{tall data.frame of annual use data.}

\item{masterspecieslist}{data.frame containing the full species list.}
}
\value{
data.frame of species-based indicator data.
}
\description{
Combine species-level indicator information into one data.frame.
}
