% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_cover_lentic.R
\name{pct_cover_lentic}
\alias{pct_cover_lentic}
\title{Calculate percent cover of grouping variables categories.}
\usage{
pct_cover_lentic(lpi_tall, tall = FALSE, hit = "any", by_line = FALSE, ...)
}
\arguments{
\item{lpi_tall}{A tall/long-format data frame. Use the data frame from the \code{gather_lpi_lentic()} output, or a
modified version of the \code{lpi_tall} table that adds categorical information used in cover calculations.}

\item{tall}{Logical. If TRUE then the returned data frame will be tall rather than wide and will not have
observations for non-existent values. Defaults to FALSE.}

\item{hit}{Character string. Absolute cover can be calculated from "any", "first", or "basal" hits. This
will count all pin drops with hits fitting into \code{grouping_variable} categories in specified layers and
calculate their cover relative to total pin drops. If "first" is used, only \code{"TopCanopy"} hits will be counted.
If "basal" is used, only \code{"SoilSurface"} hits will be counted. Relative cover can be calculated from "all" hits,
counting all hits of vascular species within \code{grouping_variable} categories. Defaults to "any".}

\item{by_line}{Logical. If TRUE then results will be reported further grouped by line using '\code{"LineKey"}.
Defaults to FALSE.}

\item{...}{Optional character strings. One or more variable name to calculate percent cover for, i.e.
"GrowthHabit", "Duration", "Nativity", or "WetlandIndicatorStatus".}
}
\value{
Data frame of the percent cover by plot of each category combination of the grouping variables provided.
}
\description{
Cover metric function use this underlying function to perform final cover calculation. \code{lpi_tall}
data frame needs to contain the column used to group variables with all categories to be included.

Depending on the variable, some manipulation to the lpi_tall dataframe may be required prior to feeding it into this
function. When variables have meaningful nulls, such as Wetland Indicator status, care needs to be taken in the handling
of meaningfully null classification compared with null values associated with unclassified higher order taxa. Further, if
multiple categories are to be grouped in the metric reporting, they must be grouped prior to absolute cover calculations,
since hits containing both categories should count as a single positive pin drop in absolute cover.

This function also allows for users to calculate either relative (i.e. proportion of vascular species hits of a particular
category) or absolute (i.e. proportion of LPI pin drops of a particular category) cover using the \code{hit} argument.
This argument controls how the denominator is calculated. It also changes whether duplicate hits in a single pin drop
are removed from the calculation (as in the case of absolute cover) or kept in (as with relative cover).
}
