% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Characterization.R
\name{header_build_lentic}
\alias{header_build_lentic}
\title{Header for lentic projects}
\usage{
header_build_lentic(dsn, ...)
}
\arguments{
\item{dsn}{File path to the file geodatabase where all raw data from AGOL is stored.}

\item{...}{filter expression in grepl format.}
}
\description{
Headers build from the plot characterization form to match to each entry from all tall tables.
}
