% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list_lentic}
\alias{species_list_lentic}
\title{Build a species list of all species seen during the field season.}
\usage{
species_list_lentic(
  header,
  spp_inventory_tall,
  lpi_tall,
  height_tall,
  woody_species
)
}
\arguments{
\item{header}{Source of header Rdata file}

\item{lpi_tall}{Source of lpi Rdata file}

\item{height_tall}{Source of height Rdata file}

\item{species_inventory_tall}{Source of species inventory Rdata file}

\item{species_file}{File path to species file if you want species attributes or updated species. Geodatabase or csv allowed.}

\item{...}{Filtering expression to subset the number of plots}
}
\description{
Build a species list of all species seen during the field season.
}
\examples{
# Get a list of all species occurring on a plot across methods (LPI, height, species inventory)
# This method also adds cover and height by species. Be aware that sample sizes may be insufficient to make an accurate estimate
}
