% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_QC_lentic.R, R/species_list.R
\name{species_list_check_lentic}
\alias{species_list_check_lentic}
\alias{species_QC_lentic}
\title{Check species data}
\usage{
species_list_check_lentic(dsn_tall = dsn_tall, species_list_file, ...)

species_QC_lentic(header, spp_inventory, lpi_tall, height_tall, woody_species)
}
\arguments{
\item{dsn_tall}{A file path to tall tables produced from gathering functions}

\item{species_list_file}{The file path to the species list. If contained within the dsn, then specify the layer name within the dsn.}

\item{...}{Filtering expression to subset the number of plots}

\item{header}{Source of header Rdata file}

\item{lpi_tall}{Source of lpi Rdata file}

\item{height_tall}{Source of height Rdata file}

\item{species_inventory_tall}{Source of species inventory Rdata file}

\item{species_file}{File path to species file if you want species attributes or updated species. Geodatabase or csv allowed.}
}
\description{
Functions which build from LPI, Height, and Species Inventory tall tables to perform species checks.
}
\examples{
# Get a list of all species occurring on a plot across methods (LPI, height, species inventory)
# This method also adds cover and height by species. Be aware that sample sizes may be insufficient to make an accurate estimate
}
