% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{summarize_height}
\alias{summarize_height}
\title{Calculate average herbaceous and woody height and litter and water depth.}
\usage{
summarize_height(
  height_tall,
  method = "mean",
  tall = FALSE,
  by_line = FALSE,
  omit_zero = TRUE
)
}
\arguments{
\item{height_tall}{Source of height_tall dataframe.}

\item{method}{Character string.  Indicates the type of summary to calculate, \code{"max"}, which yields the average maximum
height on the plot or \code{"mean"} which yields the mean height.}

\item{tall}{Logical. If TRUE then the returned data frame will be tall rather than wide and will not have
opbervations for non-existent values. Defaults to FALSE.}

\item{by_line}{Logical. If TRUE then results will be reported further grouped by line using 'LineKey.
Defaults to FALSE.}

\item{omit_zero}{Logical. Indicates whether zeros should be included in average height and depth calculations.
Defaults to TRUE.}
}
\value{
Data.frame of the summarized height data by plot grouped into herbaceous, litter, water, woody, or woody2 categories.
}
\description{
Calculate average herbaceous and woody height and litter and water depth.
}
