% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodystructure_use_metrics.R
\name{use_metrics}
\alias{use_metrics}
\alias{ageclass_metrics}
\alias{hummocks_metrics}
\title{Calculate use metrics from stubble height and woody data.}
\usage{
use_metrics(
  header,
  annualuse_tall,
  woody_tall,
  masterspecieslist,
  by_line = F,
  by_species = F
)

ageclass_metrics(
  header,
  woody_tall,
  masterspecieslist,
  by_line = F,
  by_species = F
)

hummocks_metrics(hummocks, by_line = F)
}
\arguments{
\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output.}

\item{annualuse_tall}{Data frame. Use the data frame from the \code{gather_annualuse()} output.}

\item{woody_tall}{Data frame. Use the data frame from the \code{gather_woodyspecies()} output.}

\item{by_line}{Logical. If TRUE then results will be reported further grouped by line using 'LineKey. Defaults to FALSE.}

\item{by_species}{Logical. If TRUE then results will be reported at the species-plot level. Defaults to FALSE.}

\item{hummocks}{Data frame. Use the data frame from the \code{gather_hummocks()} output.}
}
\value{
Data.frame of the summarized woody and annual use data by plot.
}
\description{
Calculate use metrics from stubble height and woody data.
}
