% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmetrics.R
\name{allmetrics_byspecies}
\alias{allmetrics_byspecies}
\alias{allmetrics_byplot}
\title{Combine species-level indicator information into one data.frame.}
\usage{
allmetrics_byspecies(
  header,
  spp_inventory,
  lpi_tall,
  height_tall,
  woody_tall,
  annualuse_tall,
  masterspecieslist,
  unknowncodelist
)

allmetrics_byplot(
  header,
  spp_inventory,
  lpi_tall,
  height_tall,
  woody_tall,
  annualuse_tall,
  hummocks,
  unknowncodelist,
  masterspecieslist
)
}
\arguments{
\item{header}{data.frame produced by \code{gather_header_lentic}.}

\item{spp_inventory}{data.frame produced by \code{gather_spp_inventory_lentic}.}

\item{lpi_tall}{tall data.frame of LPI detail data produced by \code{gather_lpi_lentic}}

\item{height_tall}{tall data.frame of height data.}

\item{woody_tall}{tall data.frame of woody use data.}

\item{annualuse_tall}{tall data.frame of annual use data.}

\item{masterspecieslist}{data.frame containing the full species list.}

\item{unknowncodelist}{optional dataframe. Unknown species list matching unknown codes to their duration and
Growth habit. This is used to fill in duration and growth habit for plants in LPI never identified to a
species or genus with those fields specified. If argument is unused, all unknown species without duration or
Growth Habit specified will be filtered out before being passed on to pct_cover_lentic.}

\item{hummocks}{tall data frame of hummock data.}
}
\value{
data.frame of species-based indicator data.
}
\description{
Combine species-level indicator information into one data.frame.
}
