% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationcheck.R
\name{classificationcheck}
\alias{classificationcheck}
\title{Combine species-level and plot-level indicator information into one data.frame used to automate classifications}
\usage{
classificationcheck(
  header,
  lpi_tall,
  height_tall,
  woody_tall,
  masterspecieslist,
  unknowncodes,
  soils,
  PlotChar
)
}
\arguments{
\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output.}

\item{lpi_tall}{Data frame. Use the data frame from the \code{gather_lpi_lentic()} output.}

\item{height_tall}{Data frame. Use the data frame from the \code{gather_heights_lentic()} output.}

\item{woody_tall}{Data frame. Use the data frame from the \code{gather_woodyspecies()} output.}

\item{masterspecieslist}{Data frame. The centrally managed master species list should be used.
#'@param unit String. The sampling unit by which data should be summarized. Should be \code{by_plot}, \code{by_line} or \code{by_geosurface} (for data from Lotic-Integration Plots). Defaults to \code{by_plot}.}

\item{unknowncodes}{Optional data frame. Use the data frame from the \code{gather_unknowns_lentic()} output. Unknown species list matching unknown codes to their duration and Growth habit. This is used to fill in duration and growth habit for plants in LPI never identified to a species or genus with those fields specified. If argument is unused, all unknown species without duration or Growth Habit specified will be filtered out before being passed on to \code{pct_cover_lentic()}.}

\item{spp_inventory_tall}{Data frame. Use the data frame from the \code{gather_spp_inventory_lentic()} output.}

\item{annualuse_tall}{Data frame. Use the data frame from the \code{gather_annualuse()} output.}

\item{hummocks}{Data frame. Use the data frame from the \code{gather_hummocks()} output.}

\item{gap_tall}{optional data frame. Use the data frame from the \code{gather_gap()} output.}

\item{soil_stability_tall}{Optional data frame. Use the data frame from the \code{gather_soilstab()} output.}

\item{waterqualdet}{Optional data frame. Use the data frame from the file geodatabase. Must have EvaluationID as the identifying column.}
}
\value{
Data frame of indicator data calculated by EvaluationID.
}
\description{
Functions starting with tall table outputs from gather functions to calculate metrics used to classify sites, then perform that classification using a combination of indicator data.
}
