% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_composition.R
\name{Community_Composition}
\alias{Community_Composition}
\title{Summarize Species or plant group by EvaluationID based on species list.}
\usage{
Community_Composition(SpeciesList, method = "percent", tall = F, ...)
}
\arguments{
\item{SpeciesList}{Data frame. Table of species by plot that will be summarized, including the grouping variables for which data is being summarized by. Duplicates and unwanted grouping variables should be filtered out prior to submitting to function.}

\item{method}{character string. The method used for the produced summary table. Can be "percent", "mean", or "count". Defaults to "percent".}

\item{tall}{Logical. Indicates whether output will be tall table or wide. Defaults to FALSE.}

\item{...}{Optional character strings. Grouping variables from the master species list to be used in calculating list.}
}
\value{
Data frame of summarized metrics by plot.
}
\description{
Underlying function used in community metric functions that summarizes the percent, count, or average of an indicated variable found in a provided species list. All unique combinations of the grouping variable used will be included in the final output, excluding NA or Null values.The species list provided should already be cleaned to remove duplicate species or unwanted grouping variable categories.
}
