% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_metrics.R
\name{Community_Richness}
\alias{Community_Richness}
\alias{Community_C.Value}
\alias{Community_Native}
\alias{Community_NoxiousCount}
\alias{Community_HydroNoFAC}
\alias{Community_HydroWithFAC}
\alias{Community_SGGroup}
\alias{Community_GrowthHabitSub}
\alias{Community_Duration}
\alias{Community_StabilityGrowthHabit}
\title{Summarize community composition for specified grouping variables.}
\usage{
Community_Richness(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory"
)

Community_C.Value(
  header,
  SpeciesList,
  statespecieslist,
  listtype = "speciesinventory"
)

Community_Native(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_NoxiousCount(
  header,
  SpeciesList,
  statespecieslist,
  listtype = "speciesinventory"
)

Community_HydroNoFAC(
  header,
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_HydroWithFAC(
  header,
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_SGGroup(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_GrowthHabitSub(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_Duration(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_StabilityGrowthHabit(
  SpeciesList,
  nationalspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)
}
\arguments{
\item{SpeciesList}{Data frame. Table of species by PlotID that will be summarized. Can be either species_inventory_tall produced by \code{gather_spp_inventory_lentic()} or lpi_tall produced by \code{gather_lpi_lentic()}.}

\item{nationalspecieslist}{Data frame. The centrally managed master species list should be used. The assumed structure is that each row is unique on its Symbol.}

\item{listtype}{Character string. Indicates the source of the SpeciesList provided. Can either be "speciesinventory" or"lpi". Defaults to "speciesinventory".}

\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output. Used in Noxious and Wetland Indicator calculations to specify the plot WetlandIndicatorRegion or state.}

\item{statespecieslist}{Data frame. The centrally managed master species list should be used. This dataframe should contain a unique record for each Symbol-SpeciesState combination.}

\item{method}{character string. The method used for the produced summary table. Can be "percent", "mean", or "count". Included in functions where applicable.}
}
\value{
Data frame of summary metrics by plot.
}
\description{
This group of functions prepares a species list which is then used in the \code{community_composition()} function. This can involve filtering duplicate species, combining grouping variable categories, and pulling plot information from plot headers to inform species classification. Analogous to \code{cover_metrics} functions.

Similarly to relative cover calculations, all species with unknown traits (e.g. an unknown plant identified only as Elymus sp.) must be filtered out prior to calculating community metrics. All community metric calculations may be calculated using \code{Community_Metrics()}
}
