% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_metrics.R
\name{Community_Richness}
\alias{Community_Richness}
\alias{Community_C.Value}
\alias{Community_Native}
\alias{Community_NoxiousCount}
\alias{Community_Hydrophytes}
\alias{Community_HydroFAC}
\alias{Community_GrowthHabit}
\alias{Community_Duration}
\title{Summarize community composition for specified grouping variables.}
\usage{
Community_Richness(
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory"
)

Community_C.Value(
  header,
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory"
)

Community_Native(
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_NoxiousCount(
  header,
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory"
)

Community_Hydrophytes(
  header,
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_HydroFAC(
  header,
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_GrowthHabit(
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)

Community_Duration(
  SpeciesList,
  masterspecieslist,
  listtype = "speciesinventory",
  method = "percent"
)
}
\arguments{
\item{SpeciesList}{Data frame. Table of species by PlotID that will be summarized. Can be either spp_inventory_tall
produced by \code{gather_spp_inventory_lentic()} or lpi_tall produced by \code{gather_lpi_lentic()}.}

\item{masterspecieslist}{Data frame. The centrally managed master species list should be used.}

\item{listtype}{Character string. Indicates the source of the SpeciesList provided. Can either be
"speciesinventory" or"lpi". Defaults to "speciesinventory".}

\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output. Used in Noxious and
Wetland Indicator calculations to specify the plot region or state.}

\item{method}{character string. The method used for the produced summary table. Can
be "percent", "mean", or "count". Included in functions where applicable.}
}
\value{
Data frame of summary metrics by plot.
}
\description{
This group of functions prepares a species list to be used in the embedded
\code{community_composition()}. This can involve filtering duplicate species, combining
grouping variable categories, and pulling plot information from plot headers to inform
species classification. Analogous to \code{cover_metrics} functions.
}
