% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_riparianwetland.R
\name{gather_RiparianWetland}
\alias{gather_RiparianWetland}
\alias{gather_lpi_lentic}
\alias{gather_species_inventory_lentic}
\alias{gather_unknowns_lentic}
\alias{gather_height_lentic}
\alias{gather_annualuse}
\alias{gather_woodyspecies}
\alias{gather_tree}
\alias{gather_hummocks}
\alias{gather_gap}
\alias{gather_soilstab}
\alias{gather_all_riparianwetland}
\title{Transform raw datatables from multiple sources into standard format}
\usage{
gather_lpi_lentic(dsn, source = "SDE", lr = FALSE)

gather_species_inventory_lentic(dsn, source = "SDE")

gather_unknowns_lentic(dsn, source = "SDE", familygenuslist = NULL)

gather_height_lentic(dsn, source = "SDE", lr = FALSE)

gather_annualuse(dsn, source = "SDE")

gather_woodyspecies(dsn, source = "SDE")

gather_tree(dsn, source = "SDE")

gather_hummocks(dsn, source = "SDE")

gather_gap(dsn, source = "SDE")

gather_soilstab(dsn, source = "SDE")

gather_all_riparianwetland(
  dsn,
  source = "SDE",
  lr = FALSE,
  familygenuslist = NULL,
  covariates = TRUE
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extensions) of the geodatabase containing the table of interest.For the AGOL source, a URL can be used.}

\item{source}{Character string. The source and schema of the data being analyzed. Default is SDE, but other options are AGOL and GDB. AGOL anticipates loading data from a feature service URL from online, while GDB anticipates a structure identical to the Survey123 project but from a local File Geodatabase.}

\item{lr}{Logical. Only used for LR integration data from LPI, where Geomorphic surface field should come with the data. Defaults to FALSE.}

\item{familygenuslist}{data.frame. Only required in gathering Unknown Plant form for data loaded from the online feature service. Otherwise, script expects Unknown Plants has already been corrected to fill in unknown plants with their codes.an exhaustive list of all possible family and genus names ('ScientificName'), their associated codes ('Code'), and the taxonomic level ('Level'), i.e. "Family" or "Genus".}
}
\value{
tall Data frame containing the data from a detail table from a Riparian and Wetland AIM file geodatabase.
}
\description{
This group of functions transforms data from varied formats into one uniform structure. It contains functions for transforming data from species inventory, unknown plants, LPI, LPI heights from LPI, Woody Structure, Annual Use, Hummocks, Gap, and Soil Stability tables. For LPI, gathering pivots the data from wide to long format. Gathering Unknown Plants also corrects field season data from online data. Most other transformations merely combine data from parent and child tables into a single table to be used in analysis and ensure consistent column names across sources.
}
