% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdbconversion.R
\name{gdbconversion}
\alias{gdbconversion}
\title{Convert FGDB into separate .csv files}
\usage{
gdbconversion(dsn, RawDataFolder)
}
\arguments{
\item{dsn}{location of the File Geodatabase}

\item{RawDataFolder}{location where all csvs will be stored.}
}
\value{
Dataframe of specified perce cover of different categories by plot.
}
\description{
Function converts all separate tables found in a FGDB into separate csv files to be
more accessible to others. Some schema changes are accomplished to ensure output is
interpretable outside of a relational geodatabase format.
}
