% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_build.R
\name{header_build_lentic}
\alias{header_build_lentic}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
header_build_lentic(dsn, ...)
}
\arguments{
\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}
}
\value{
A \code{tbl} of header information on each evaluated plot.
}
\description{
Build AIM Indicators Tables and Feature Classes
}
