% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_build.R
\name{header_build_lentic}
\alias{header_build_lentic}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
header_build_lentic(dsn, ...)
}
\arguments{
\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}

\item{header}{Dataframe. Plot header containing plot metadata}

\item{source}{String. Specifies data source, \code{"AIM", "LMF"}}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes
}
