% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_metrics.R
\name{height_metrics}
\alias{height_metrics}
\title{Calculate average herbaceous and woody height and litter and water depth.}
\usage{
height_metrics(
  height_tall,
  masterspecieslist,
  unknowncodes,
  method = "mean",
  by_line = FALSE,
  omit_zero = TRUE,
  by_species = FALSE
)
}
\arguments{
\item{height_tall}{Data frame. Use the data frame from the \code{gather_height_lentic()} output.}

\item{masterspecieslist}{Data frame. The centrally managed master species list should be used.}

\item{unknowncodes}{Optional data frame. Use the data frame from the \code{gather_unknowns_lentic()} output. Unknown species list matching unknown codes to their duration and Growth habit. This is used to fill in growth habit for plants in LPI never identified to a species or genus with those fields specified. This information is used to filter out herbaceous height measurements of non-herbaceous species and woody height measurements of non-woody species. If not included, unknown growth habit measurements will be maintained in the final height calculations.}

\item{method}{Character string.  Indicates the type of summary to calculate, \code{"max"}, which yields the average maximum height on the plot or \code{"mean"} which yields the mean height.}

\item{by_line}{Logical. If TRUE then results will be reported further grouped by line using 'LineKey. Defaults to FALSE.}

\item{omit_zero}{Logical. Indicates whether zeros should be included in average height and depth calculations. Defaults to TRUE.}

\item{by_species}{Logical. If TRUE, then results will  reported by species-plot rather than at the plot or transect level. Defaults to FALSE.}
}
\value{
Data frame of the summarized height data by plot grouped into herbaceous, litter, water, or woody categories.
}
\description{
Calculate average herbaceous and woody height and litter and water depth.
}
