% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_byspecies.R
\name{metrics_byspecies}
\alias{metrics_byspecies}
\title{Combine species-level indicator information into one data.frame.}
\usage{
metrics_byspecies(
  header,
  spp_inventory,
  lpi_tall,
  height_tall,
  woody_tall,
  annualuse_tall,
  masterspecieslist
)
}
\arguments{
\item{header}{source of header data frame.}

\item{spp_inventory}{data.frame produced by \code{gather_spp_inventory_lentic}.}

\item{lpi_tall}{source of lpi_tall data frame.}

\item{height_tall}{Source of height_tall dataframe.}

\item{woody_tall}{tall data frame of woody use data.}

\item{annualuse_tall}{tall data frame of annual use data.}

\item{masterspecieslist}{Character string. Full file path (including extension) to the file
containing the species list.}
}
\value{
Data.frame of species-based indicator data.
}
\description{
Combine species-level indicator information into one data.frame.
}
