% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_cover_lentic.R
\name{pct_cover_lentic}
\alias{pct_cover_lentic}
\title{Function used to calculate the percent cover by different grouping variables.}
\usage{
pct_cover_lentic(lpi_tall, tall = FALSE, hit = "any", by_line = FALSE, ...)
}
\arguments{
\item{lpi_tall}{source of lpi_tall data frame.}

\item{tall}{Logical. If TRUE then the returned data frame will be tall rather than wide and will not have
opbervations for non-existent values. Defaults to FALSE.}

\item{hit}{Character string. Absolute cover can be calculated from "any", "first", or "basal" hits. This
will count all pin drops with hits fitting into \code{grouping_variable} categories in specified layers and
calculate their cover relative to total pin drops. If "first" is used, only TopCanopy hits will be counted.
If "basal" is used, only SoilSurface hits will be counted. Relative cover can be calculated from "all" hits,
counting all hits of vascular species within \code{grouping_variable} categories. Defaults to "any".}

\item{by_line}{Logical. If TRUE then results will be reported further grouped by line using 'LineKey.
Defaults to FALSE.}

\item{...}{Optional character strings. One or more variable name to calculate percent cover for, i.e.
"GrowthHabit", "Duration", "Nativity", or "WetlandIndicatorStatus".}

\item{masterspecieslist}{Character string. Full file path (including extension) to the file
containing the species list.}
}
\description{
Function used to calculate the percent cover by different grouping variables.
}
