% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_QC_lentic.R
\name{species_list_check_lentic}
\alias{species_list_check_lentic}
\title{Check species data}
\usage{
species_list_check_lentic(dsn_tall = dsn_tall, species_list_file, ...)
}
\arguments{
\item{dsn_tall}{A file path to tall tables produced from gathering functions}

\item{species_list_file}{The file path to the species list. If contained within the dsn, then specify the layer name within the dsn.}
}
\description{
Functions which build from LPI, Height, and Species Inventory tall tables to perform species checks.
}
