% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodystructure_use_metrics.R
\name{use_metrics}
\alias{use_metrics}
\alias{ageclass_metrics}
\alias{SGConifer_metrics}
\alias{hummocks_metrics}
\title{Calculate use metrics from stubble height and woody data.}
\usage{
use_metrics(
  header,
  annualuse_tall,
  woody_tall,
  masterspecieslist,
  unit = "by_plot",
  by_species = F
)

ageclass_metrics(
  header,
  woody_tall,
  tree_tall = NULL,
  masterspecieslist,
  unit = "by_plot",
  by_species = F
)

SGConifer_metrics(tree_tall, masterspecieslist, unit = "by_plot")

hummocks_metrics(hummocks, unit = "by_plot")
}
\arguments{
\item{header}{Data frame. Use the data frame from the \code{header_build_lentic()} output.}

\item{annualuse_tall}{Data frame. Use the data frame from the \code{gather_annualuse()} output.}

\item{woody_tall}{Data frame. Use the data frame from the \code{gather_woodyspecies()} output.}

\item{unit}{String. The sampling unit by which data should be summarized. Should be \code{by_plot}, or \code{by_line}. \code{by_geosurface} is not an option for this calculation. Defaults to \code{by_plot}.}

\item{by_species}{Logical. If TRUE then results will be reported at the species-plot level. Defaults to FALSE.}

\item{tree_tall}{Optional data frame. Use the data frame from the \code{gather_tree()} output. For data from years prior to 2023, this data.frame will not exist.}

\item{hummocks}{Data frame. Use the data frame from the \code{gather_hummocks()} output.}
}
\value{
Data.frame of the summarized woody and annual use data by plot.
}
\description{
Calculate use metrics from stubble height and woody data.
}
